LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/fetch/metadata/generated/css-font-face.sub.tentative-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY

FAIL sec-fetch-site - Not sent to non-trustworthy same-origin destination assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - Not sent to non-trustworthy cross-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-mode - Not sent to non-trustworthy same-origin destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-mode - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-mode - Not sent to non-trustworthy cross-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-dest - Not sent to non-trustworthy same-origin destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-dest - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-dest - Not sent to non-trustworthy cross-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-user - Not sent to non-trustworthy same-origin destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-user - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-user - Not sent to non-trustworthy cross-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - HTTPS downgrade (header not sent) promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - HTTPS upgrade promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - HTTPS downgrade-upgrade promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
